有生之年.jpg

+++ src/main/java/me/skymc/taboolib/particle/EffLib.java
@@ -1176,7 +1176,7 @@ public enum EffLib {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) red / 255F;
+			return red / 255F;
 		}
 
 		/**
@@ -1186,7 +1186,7 @@ public enum EffLib {
 		 */
 		@Override
 		public float getValueY() {
-			return (float) green / 255F;
+			return green / 255F;
 		}
 
 		/**
@@ -1196,7 +1196,7 @@ public enum EffLib {
 		 */
 		@Override
 		public float getValueZ() {
-			return (float) blue / 255F;
+			return blue / 255F;
 		}
 	}
 
@@ -1234,7 +1234,7 @@ public enum EffLib {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) note / 24F;
+			return note / 24F;
 		}
 
 		/**
